requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
pymongo>=4.7.0
dnspython>=2.6.0
pyjwt>=2.8.0
pydantic[email]>=2.7.0
python-multipart>=0.0.9
python-dotenv>=1.0.0
bcrypt>=4.1.0
slowapi>=0.1.9
circuitbreaker>=2.0.0
pytest>=8.1.0
pytest-asyncio>=0.23.0
httpx>=0.27.0
mongomock>=4.1.2